Added support for github pages
Added configuration ins package.json and yarn.lock for the git pages.
Now user can visit the gitHub site to see a live demo of the project
from any supported device.
This commit is verfied and all the changes done are working and the site
is also hosted properly without any issues.