Updated to reset the local storage when reseting
Added a line which is used to reset the local storage when it is
called.
Updated it because of some glitches occuring:
If a user was reseting and leaving the game then the game would start
before the reset as the local storage was not updated.
Now fixed the issue