Removed the changed property from the Node
Removed the changed property which was to track the node if it is
changed by the user or not.
But this same can be achieved with the help of the isModifiable property
which eleminates the need of the "changed" property.

This change removes overhead from the code which makes the code more
maintable for future and clean.

-> No issue faced till now because of this change.