Added solve feature
A button was added to the UI which is used for solving the grid, and
solve function was implemented to solve which is to be called when we
click on the button.

Users Now can view the final solution of the sudoko puzzle, At any stage
of the puzzle[If the sudoko contains more than one solution].
Users are also prompted with a message if no solution for the current
user grid is present.

Features to add:
1. We can implement a feature which can be used to show step by step
   solution or a hint for the user.