Added isHinted attribute to the node
Now Node has a attribute called isHinted which makes the node orange is
the Node value was given by the computer and not by the user when using
the solve or hint button[Hint button is not available for not, will be
added in next commit]

Users now can see which node were filled by them and also which nodes
where filled by the solve method, which makes the user to how much
progress they made.

Things to do:
Add a hint button to only get value of one node, so that users can play
the game after the get a hint.