Added Random hint in place of linear hint
Now handleHint function gives a random hint in place of the linear hint
as in previous.
This change will help the users to be immersed in solving the sudoko,
since linear hint broke the flow of player.
One can try both the version using the git and repond to it.