Added a error check for empty solutions
Added a error check for the hint function, if the board is solved and
no further hint is possible.

In further commits, we have to find a nice way to end the game and also
replace this with other!
Users now will not face any issue regarding the hint crash after the
whole board is solved.