Updated the code so that user cannot change the hint
Added a check at solve and hint for the hinted node so that users cannot
change the hinted node after it is displayed.

Users now cannot change the value of a hint, since it is a hint which
means that the value is correct and not needs to be changed by the user.

Some Unknown error may occur in this commit because of the edge cases!