Added hook for connecting localstorage and useState
Now one can use the custom hook useLocalStorage to store information in
local storage and aswell as in the state.

It will be very helpful to keep the code clean when add more conditions
for score and winning.
For now we are using useLocalStorage at:
1. Storing the Starting Grid
2. Storing the Current player Grid
3. Storing the current choice of the user

This change is a big change as it will help to keep the code clean in
the future, as we will not be manually configuring.

Note: One can look at the previous commit code t osee how grid was
stored and how we connect it with local storage which made the code not
clean and it was not scalable.