Grid Generation Algorithm Changed to give unique solution
Changed the algorithm for removing the solved grid cells to get unique
solution grid, before this update a sudoku grid generated may have more
than one solution, which was a problem as all Real sudoku grid has only
one unique solution.

Now with the help of the new algorithm we first check the solution count
of the grid when removing the cell and if count is 1 the we make the
cell empty else we keep the cell unchanged and head to the next cell.

Now users can enjoy the full feature of sudoku and no Invalid grid will
be generated.

Note: When making a new grid, sometimes it take more than 1.5s to 2s
which should be managed by improving the algorithm or by introducing a
loader for the game.