Added Choice for difficulty on New Game
Added a Component called DifficultySelectionModal for chosing
difficulty of the sudoku game, Here difficulty is taken proportional to
the number of empty cells.

Note:
Try to implelment the hard difficulty with 60 empty cells without
lagging, for now we have taken 50 empty cells!
Tried implementing for 60 empty cells a lot,but it is a very CPU
intensive task which makes the react website lag.
Will try to implement this in future !