Added Difficulty mode
Added difficulty mode and also made global variable defining the modes
and how many empty cells each mode have.

Users now can view the mode they are playing by visiting the details
page.

Removed useEffect for checking if grid is null or not ,and now we check
it after each render as it is more clean way to do it than the previous
way